Remodel import and directory structure

+++ src/fastsci/core/matrix/main.h
@@ -7,8 +7,8 @@ extern "C" {
 #endif
 
 void matmul(const double* A, const double* B, double* C, size_t M, size_t N, size_t K);
-void add_scalar(const double* a, double scalar, double* out, int size);
-void mul_scalar(const double* a, double scalar, double* out, int size);
+void add_scalar(const double* a, double scalar, double* out, size_t size);
+void mul_scalar(const double* a, double scalar, double* out, size_t size);
 
 #ifdef __cplusplus
 }
